employee seeder

+++ database/migrations/2019_07_18_055115_create_employees_table.php
@@ -16,7 +16,7 @@ class CreateEmployeesTable extends Migration
         Schema::create('employees', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->integer('nik')->unique();
+            $table->bigInteger('nik')->unique();
             $table->string('name');
             $table->string('ktp')->unique();
             $table->string('npwp')->nullable();
